ai-service/requirements.txt
flask==3.0.0
openai-whisper
gunicorn==22.0.0
